sending http requests to our api routes
