+++ 26---A-Pretty-Deep-Dive-Introduction-to-Next.js/06-onwards-to-a-bigger-project-starter/pages/new-meetup/index.js
@@ -4,8 +4,16 @@ import NewMeetUpForm from "../../components/meetups/NewMeetupForm";
 
 function NewMeetupPage() {
 
-    function addMeetupHandler(enteredMeetupData) {
-        console.log(enteredMeetupData)
+    async function addMeetupHandler(enteredMeetupData) {
+        const response = await fetch("/api/new-meetup", {
+            method: "POST",
+            body: JSON.stringify(enteredMeetupData),
+            headers: {
+                "Content-Type": "application/json"
+            }
+        });
+        const data = await response.json();
+        console.log(data);
     }
 
     return <NewMeetUpForm onAddMeetup={addMeetupHandler}/>
